+++ src/Space.hs
@@ -25,14 +25,14 @@ import Common ((~==))
 import Matrix (SpaceMatrix, fromTuple4)
 
 class SpaceElement a where
-    getX :: a -> Float
-    getY :: a -> Float
-    getZ :: a -> Float
+    getX :: a -> Double
+    getY :: a -> Double
+    getZ :: a -> Double
 
 eq :: (SpaceElement a) => a -> a -> Bool
 e1 `eq` e2 = (getX e1 ~== getX e2) && (getY e1 ~== getY e2) && (getZ e1 ~== getZ e2)
 
-data Point = Point { getPointX :: Float, getPointY :: Float, getPointZ :: Float } deriving (Show)
+data Point = Point { getPointX :: Double, getPointY :: Double, getPointZ :: Double } deriving (Show)
 
 instance SpaceElement Point where
     getX = getPointX
@@ -42,7 +42,7 @@ instance SpaceElement Point where
 instance Eq Point where
     (==) = eq
 
-data Vector = Vector { vectorX :: Float, vectorY :: Float, vectorZ :: Float } deriving (Show)
+data Vector = Vector { vectorX :: Double, vectorY :: Double, vectorZ :: Double } deriving (Show)
 
 instance SpaceElement Vector where
     getX = vectorX
@@ -73,13 +73,13 @@ Point x1 y1 z1 `subtractVectorP` Vector x2 y2 z2 = Point (x1 - x2) (y1 - y2) (z1
 subtractVectorV :: Vector -> Vector -> Vector
 Vector x1 y1 z1 `subtractVectorV` Vector x2 y2 z2 = Vector (x1 - x2) (y1 - y2) (z1 - z2)
 
-multiplyVector :: Vector -> Float -> Vector
+multiplyVector :: Vector -> Double -> Vector
 Vector x y z `multiplyVector` s = Vector (x * s) (y * s) (z * s)
 
-divideVector :: Vector -> Float -> Vector
+divideVector :: Vector -> Double -> Vector
 Vector x y z `divideVector` d = Vector (x / d) (y / d) (z / d)
 
-magnitude :: Vector -> Float
+magnitude :: Vector -> Double
 magnitude (Vector x y z) = sqrt (x ** 2 + y ** 2 + z ** 2)
 
 normalize :: Vector -> Vector
@@ -87,7 +87,7 @@ normalize (Vector x y z) =
     let m = magnitude (Vector x y z)
     in Vector (x / m) (y / m) (z / m)
 
-dot :: Vector -> Vector -> Float
+dot :: Vector -> Vector -> Double
 dot (Vector x1 y1 z1) (Vector x2 y2 z2) = x1 * x2 + y1 * y2 + z1 * z2
 
 cross :: Vector -> Vector -> Vector
